LInear correction

+++ Project/src/main/java/Threads/CalculateLinearVariablesThread.java
@@ -15,9 +15,9 @@ public class CalculateLinearVariablesThread extends Thread{
 	private ElasticSearchService linearDataClient ;
 	private Map<String,Object> trainingDataRecord , proteinAsMap;
 	private HammingCalculation m_hammingCalculation;
-	private long firstProteinIndex , secondProteinIndex ,fragmentHammingDistance , contextHammingDistance; 
+	private int firstProteinIndex , secondProteinIndex ,fragmentHammingDistance , contextHammingDistance; 
 	private Protein firstProtein , secondProtein;
-	private long firstProteinFragment , secondProteinFragment;
+	private int firstProteinFragment , secondProteinFragment;
 	
 	public CalculateLinearVariablesThread( ElasticSearchService trainingDataClient, ElasticSearchService proteinsDataClient, 
 											ElasticSearchService linearDataClient) {
@@ -50,12 +50,12 @@ public class CalculateLinearVariablesThread extends Thread{
 	{
 		trainingDataRecord = trainingDataClient.get((int)index);
 		
-		firstProteinIndex = (long) trainingDataRecord.get("firstProteinIndex");
-		secondProteinIndex = (long) trainingDataRecord.get("secondProteinIndex");
+		firstProteinIndex = (int) trainingDataRecord.get("firstProteinIndex");
+		secondProteinIndex = (int) trainingDataRecord.get("secondProteinIndex");
 		firstProteinFragment  = (int) trainingDataRecord.get("firstFragmentIndex");
 		secondProteinFragment = (int) trainingDataRecord.get("secondFragmentIndex");
 		
-		fragmentHammingDistance = (long)trainingDataRecord.get("HammingDistance");
+		fragmentHammingDistance = (int)trainingDataRecord.get("HammingDistance");
 		
 		firstProtein = getProteinFromDB(firstProteinIndex) ; 
 		secondProtein = getProteinFromDB(secondProteinIndex);
@@ -76,7 +76,7 @@ public class CalculateLinearVariablesThread extends Thread{
 	
 	
 	
-	private Protein getProteinFromDB(long index )
+	private Protein getProteinFromDB(int index )
 	{
 		proteinAsMap = proteinsDataClient.getProtein(index);
